refactor: 森詳細page.tsxのヘルパー4関数をforestProfileUtils.tsへ分離
App Routerのpage.tsxから任意named export（isValidForestProfileId /
normalizeDreamProfilesResponse / normalizeProfileDreamsResponse /
findActiveForestProfile）を兄弟モジュールへ移動し、pageはdefault export
のみにする。

経緯: 6月時点のNextでは「page.tsxの任意named exportでnext buildの型
チェックが失敗する」事象が記録されていたが、現行Next 16.2.7では再現
しない（main素の状態でbuild成功・生成型込みtscもエラーなしを確認済み）。
そのため本変更はビルド修正ではなく、Next側制約の再厳格化に備えた
規約整理＋森画面刷新(#4)の前捌きとして行う。ロジック変更なし。

検証: ForestProfilePage.test 6/6緑（importを新モジュールへ変更）、
Jest全体 387/387緑、next build成功。
